requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
autogen-agentchat==0.5.6
autogen-core==0.5.6
autogen-ext==0.5.6
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
fonttools==4.57.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
jiter==0.9.0
jsonref==1.1.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
ollama==0.4.8
openai==1.77.0
opentelemetry-api==1.32.1
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
